requirements.txt
astroid==2.15.0
black==23.1.0
click==8.1.3
dill==0.3.6
duckdb==0.7.1
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.1.1
pyarrow==11.0.0
pycodestyle==2.10.0
pygame==2.3.0
pylance==0.3.17
pylint==2.17.1
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
-e git+ssh://git@github.com/cpsc-spring-2023/cpsc-386-04-scene-EPW80.git@ad50ba9c83ed7b0baf89ffba40b0bd441f4bfe4f#egg=videogame&subdirectory=videogame
wrapt==1.15.0
